card save시 workList찾는 로직수정, [workList,card]entity 수정

+++ src/main/java/com/toy/trelloapi/domain/repository/WorkListRepository.java
@@ -10,5 +10,5 @@ import java.util.List;
 @Repository
 public interface WorkListRepository extends JpaRepository<WorkList, Long>, QueryByExampleExecutor<WorkList>, WorkListForQueryDsl {
     List<WorkList> findAllBy();
-    WorkList findByWorkListId(Integer id);
+    WorkList findByWorkListId(Long id);
 }
